[supernova] extract FilterPills into own component
